requirements.txt
Django
selenium
Markdown
openai
django-filter
djangorestframework